rag/requirements-indexer.txt
functions-framework==3.5.0
google-cloud-firestore==2.16.0
google-cloud-aiplatform==1.35.0
langchain==0.1.0
langchain-google-vertexai==0.1.0
langchain-google-firestore==0.1.0
